network: split TLS and generic/TCP stream support
